Change "Job Title" to "Job Title or Team"
Rename the "Job Title" field to "Job Title or Team" for those who wish
to de-emphasise hierarchy whilst maintaining the work area. Also makes
this field optional.

Also changes the labels on the form that rather than optional fields
being marked with "(Optional)", mandatory fields are marked with "*".
Also removes unnecessary colons from the labels.

Also explicitly sets the colour of the "This email may" footer to gray,
as the "Copy signature" function didn't seem to be capturing the colour
inherited from the body correctly.